Extracted render time function to utils module
